Fieldset NG: the content box should be able to be an OOF container
OOF descendants of a fieldset should not be contained by the fieldset
because they should be scrolled together with the fieldset contents.
This CL makes an anonymous fieldset content box an OOF container by
adding |filter: brightness(100%)| or |position: relative| if the
fieldset can contain OOF objects.

We need to add a special handling for OOF descendants of a LEGEND to
FindAncestorByPredicate() because such OOF descendants are contained
by fieldsets in NG fragment trees.  Without this change, we had some
CHECK failures.

Bug: 1153042
Change-Id: I993922cfd373a9e0558bc50edb009680f0f1244a